+++ CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 Types of changes: Added, Changed, Deprecated, Removed, Fixed, Security
 
+## [Unreleased]
+
+### Changed
+
+- Focus mode enhancements, option to hide gauges
+
 ## [0.12.1] - 2026-03-22
 
 ### Changed